requirements.txt
aiofiles==24.1.0
aiosmtplib==4.0.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.1
billiard==4.2.1
boto3==1.39.15
botocore==1.39.15
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
colorlog==6.9.0
cssbeautifier==1.15.4
djlint==1.36.4
dnspython==2.7.0
editorconfig==0.17.1
elastic-transport==8.17.1
elasticsearch==8.17.2
email-validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
flower==2.0.1
geomet==0.2.1.post1
googleapis-common-protos==1.70.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
jmespath==1.0.1
jsbeautifier==1.15.4
json5==0.12.0
kombu==5.5.4
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-asgi==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-instrumentation-logging==0.56b0
opentelemetry-propagator-jaeger==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.56b0
orjson==3.11.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pygments==2.19.2
pyjwt==2.10.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.8
rignore==0.6.4
ruff==0.12.5
s3transfer==0.13.1
sentry-sdk==2.33.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
text-unidecode==1.3
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
zipp==3.23.0
